app/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pandas 
numpy 
matplotlib
seaborn